+++ rhio/src/lib.rs
@@ -1,7 +1,6 @@
-#[allow(dead_code, unused, unused_imports)]
 mod blobs;
 pub mod config;
-pub mod http_server;
+pub mod health;
 mod nats;
 mod network;
 mod node;